Can now add notes to anything. It appears to work OK. Haven't tested adding notes to text block, because that would only work if the text block were not writable.
Remaining problems: (1) Notes scale with parent image, which is not desired.
This is actually interesting, because the p0 of the note should scale with
the image, so it is a nontrivial issue.
(2) For whatever reason releasing the Shift key sends a KeyRelease with key
"-1" and so the modifier never gets dropped. This is also nontrivial, because
we only learn of the modifier release at the next key press. I have to
rethink my keyboard modifier model a little bit.