Bumped version to 1.2.21
Improved handling of words like "sum" and "sharp" in math mode: These are
not longer automatically substituted to "∑" and "♯" except when preceded
by an explicit backslash.

The word "aardvark" (and, more usefully, non-english words starting with "aa")
can now be typed in math mode as-is. (Previously, italics exceptions for the
word "a" would gobble up one of the "a"s.)

Better support for double subscripts/superscripts: characters "+", "-",
"=", "(", ")" are now handled, as are multidigit numbers. Unfortunately,
there are no Unicode subscript/superscript "." and "," glyphs.