Improved insertion of closing characters
Normally, when inserting characters at the ending edge of a marked up
region, NotedELN places the new character outside of the region. Now,
closing characters (such as "]" and "”") are placed inside the region
if a corresponding opening character exists in the region.

It is still not easy to, e.g., place superscripted parentheses around
"2" in "x²" after the fact, but at least half the problem is solved.
For instance, replacing parentheses by brackets in "e^(x + y)" is now
easier.